requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
django-cors-headers>=4.4,<4.5
drf-spectacular>=0.27,<0.28
